requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pyjwt==2.8.0
cryptography>=42.0.0